Add files via upload
When opening the file in Excel, you need to 'allow macros'. Then you can use the green button to load in and convert your encrypted data.txt file (from the Power Predictor SD card). And the Excel file should do the rest. Not as pretty as the original web-app, but you can now access the wind speed data.